% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aleza}
\alias{aleza}
\title{Aleza}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 2 columns.
}
\usage{
aleza
}
\description{
A polygon of the gross tenure area of the \href{https://alrf.unbc.ca}{Aleza Lake Research Forest}. For simplicity the park in the middle is not included.
}
\examples{
aleza 

plot(aleza, main = "Aleza Lake Research Forest")
}
\keyword{datasets}
