% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baselayers}
\alias{baselayers}
\title{Baselayers}
\format{
An object of class \code{data.frame} with 8 rows and 2 columns.
}
\source{
Data layers from the BC Data Catalog.  see the package "bcdata"
}
\usage{
baselayers
}
\description{
A dataframe of BC Data Catalogue layers to queried.  Includes a human readable name and the BC Data ID.
}
\keyword{datasets}
