% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bc}
\alias{bc}
\title{BC: British Columbia}
\format{
An object of class \code{sf} (inherits from \code{bcdc_sf}, \code{tbl_df}, \code{tbl}, \code{data.frame}) with 151 rows and 1 columns.
}
\source{
\code{bcmaps::bc_bound()} or https://github.com/bcgov/bcmaps/
}
\usage{
bc
}
\description{
An 'sf' polygon of the province of British Columbia, Canada
}
\keyword{datasets}
