% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fhd.R
\name{calc_fhd}
\alias{calc_fhd}
\title{foliar height diversity}
\usage{
calc_fhd(m, method)
}
\arguments{
\item{m}{matrix of light adjusted vai values.}

\item{method}{"MH" is MacArthur-Horn and "Bohrer" is the Bohrer method}
}
\value{
foliage height diveristy
}
\description{
\code{calc_fhd} calculates foliage height diversity
}
\details{
The \code{calc_fhd} function calculates foliage height diversity, where a version of the Shannon-Weiner
diversity index is applied to through canopy measures of gap fraction
}
\examples{

calc_fhd(pcl_vai, method = "Bohrer")

}
\keyword{statisitcs}
