% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gap_fraction.R
\name{calc_gap_fraction}
\alias{calc_gap_fraction}
\title{Calculate gap fraction}
\usage{
calc_gap_fraction(m)
}
\arguments{
\item{m}{the matrix of adjusted values}
}
\description{
\code{calc_gap_fraction} imports and processes a single PCL transect.
}
\details{
This is a specific function that works using the adjusted matrix to calculate gap fraction through the canopy.
}
\examples{
calc_gap_fraction()

}
\keyword{fraction}
\keyword{gap}

