% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gini.R
\name{calc_gini}
\alias{calc_gini}
\title{Gini coefficient}
\usage{
calc_gini(m, method)
}
\arguments{
\item{m}{matrix of light adjusted vai values.}

\item{method}{"MH" is MacArthur-Horn and "Bohrer" is the Bohrer method}
}
\value{
gini coefficient
}
\description{
\code{calc_gini} calculates gini coefficient
}
\details{
The \code{calc_gini} function calculates gini coefficient, a varialbe the describes dissimilarity
}
\examples{

calc_gini(pcl_vai, method = "Bohrer")

}
\keyword{statisitcs}
