% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_intensity.R
\name{calc_intensity}
\alias{calc_intensity}
\title{Intensity Statistics}
\usage{
calc_intensity(df, filename)
}
\arguments{
\item{df}{data frame of uncorrected PCL data}

\item{filename}{name of file currently being processed}
}
\value{
statisics on the intensity data
}
\description{
\code{calc_intensity} calcualtes statisitcs from the intensity column of the PCL data
}
\details{
The \code{calc_intensity} function calculates statistics about the intensity
data in the PCL data, including min, max, sd, mean, median.
}
\examples{

intensity_stats <- calc_intensity(pcl_adjusted, filename = "UVA")

}
\keyword{statisitcs}
