% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rugosity.R
\name{calc_rugosity}
\alias{calc_rugosity}
\title{Calculate rugosity and other higher level complexity metrics}
\usage{
calc_rugosity(df, m, filename)
}
\arguments{
\item{df}{is a data frame of adjusted pcl data.}

\item{m}{matrix of light adjusted vai values.}

\item{filename}{the name of the file currently being processed.}
}
\value{
a slew of metrics...need to fully outline later
}
\description{
\code{calc_rugosity} imports and processes a single PCL transect.
}
\details{
This is a specific function calculates canopy rugosity and other metrics
}
\examples{
# Calculates metrics of canopy structural complexity.
\dontrun{calc_rugosity()
}
}
\keyword{rugosity}
