% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rumple.R
\name{calc_rumple}
\alias{calc_rumple}
\title{Calculates rumple}
\usage{
calc_rumple(df)
}
\arguments{
\item{df}{data frame of processed pcl data}
}
\value{
rumple for the canopy based on 2-D transect
}
\description{
\code{calc_rumple} imports and processes a single PCL transect.
}
\details{
This is a specific function that works using the input of a data directory
and a filename in that directory.Future versions of this will allow for direct
input of file into command so there will be no need to specify both directory and file.
}
\examples{
calc_rumple()

\dontrun{

}
}
\keyword{rumple}
