% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tls_mean_leaf_ht.R
\name{calc_tls_mean_leaf_ht}
\alias{calc_tls_mean_leaf_ht}
\title{Process single PCL transects.}
\usage{
calc_tls_mean_leaf_ht(m)
}
\arguments{
\item{m}{the vai matrix}
}
\value{
adds columns to the matrix of height.bin
}
\description{
\code{calc_tls_mean_leaf_ht} used in process_tls to calculate mean leaf height from tls slife
}
\details{
This function derives mean leaf height from x, z vai from TLS data.
}
\examples{

# with designated file
\dontrun{process_pcl("pcl_data.csv", marker.spacing = 10, user_height = 1.05, max.vai = 8)
}

}
\keyword{processing}
\keyword{tls}
