% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_vai.R
\name{calc_vai}
\alias{calc_vai}
\title{Calculate vegetation area index (VAI) from normalized PCL data matrix}
\usage{
calc_vai(df)
}
\arguments{
\item{df}{data frame of pcl data that has been corrected for light extinction}
}
\value{
a matrix of vai by x, z in the canopy
}
\description{
\code{calc_vai} imports and processes a single PCL transect.
}
\examples{
calc_vai()
\dontrun{

}
}
\keyword{vai}

