% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_variables.R
\name{combine_variables}
\alias{combine_variables}
\title{Calculate rugosity and other higher level complexity metrics}
\usage{
combine_variables(variable.list, csc.metrics, rumple, clumping.index)
}
\arguments{
\item{variable.list}{a list of variables from calc_rugosity}

\item{csc.metrics}{a list of metrics from the csc_metrics function}

\item{rumple}{rumple value from calc_rumple}

\item{clumping.index}{clumping index value}
}
\value{
a concatenated list of csc variables
}
\description{
\code{combine_variables} imports and processes a single PCL transect.
}
\details{
This is a specific function combines variables together to a coherent list.
}
\examples{
combine_Variables()

}
\keyword{csc}

