% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csc_metrics.R
\name{csc_metrics}
\alias{csc_metrics}
\title{CSC metrics basic}
\usage{
csc_metrics(df, filename)
}
\arguments{
\item{df}{data frame of light saturation corrected PCL data}

\item{filename}{name of file currently being processed}
}
\value{
slew of canopy structure metrics
}
\description{
\code{csc_metrics} imports and processes a single PCL transect.
}
\details{
This is a specific function that works using the input of a data directory
and a filename in that directory.Future versions of this will allow for direct
input of file into command so there will be no need to specify both directory and file.
}
\examples{

\dontrun{

}
}
\keyword{csc,}
\keyword{rugosity}
\keyword{structure,}

