% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etp.R
\name{etp}
\alias{etp}
\title{Calculate Potential Evapotranspiration using Thornthwaite method}
\usage{
etp(date, t_med, lat, i = NULL, a = NULL)
}
\description{
Calculate Potential Evapotranspiration using Thornthwaite method
}
