% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-reg.R
\name{fill_reg}
\alias{fill_reg}
\title{Fill gap with linear regression}
\usage{
fill_reg(.data, id, key, lon, lat, value, prop_missing = 50,
  min_coverage = 80, max_iter = 10, min_r2 = 0.8)
}
\description{
Fill gap with linear regression
}
