% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_aoi.R
\name{fm_aoi}
\alias{fm_aoi}
\title{Create an area of interest}
\usage{
fm_aoi(
  xmin = -139,
  ymin = 49,
  xmax = -114,
  ymax = 60,
  basemap = "GeoportailFrance.orthos"
)
}
\arguments{
\item{xmin, xmax}{longitude minimum and maximum extents (negative for west)}

\item{ymin, ymax}{latitude minimum and maximum extents in degrees (negative for south)}

\item{basemap}{a Leaflet basemap}
}
\value{
Returns an sf object.
}
\description{
Creates an area of interest interactively.  Simply call the function
}
\examples{
aoi <- fm_aoi(xmin = -139, ymin = 49, xmax = -114, ymax = 60)


}
