% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_collect_basemap.R
\name{fm_collect_basemap}
\alias{fm_collect_basemap}
\title{Collect Basemap Data from BC}
\usage{
fm_collect_basemap(aoi, outdir = "./out", type = "gpkg")
}
\arguments{
\item{aoi}{A \code{sf} polygon.}

\item{outdir}{A filepath to where the data should be saved. Either relative
to the working directory or an absolute path.}

\item{type}{desired spatial data structure.  Geopackage (gpkg) is the default.}
}
\value{
Spatial data is saved to the output directory.  A data.frame is returned indicating what data was received.
}
\description{
Collects essential transportation and water related layers for an area of
interest from \href{https://catalogue.data.gov.bc.ca}{BC Data}.  This function is
a wrapper for \code{bcdata}.
}
\examples{
fm_collect_basemap(aoi = fm_aoi(), outdir = "./out", type = "gpkg")
}
