% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_gpx_export.R
\name{fm_export_gpx}
\alias{fm_export_gpx}
\title{Export Spatial Data to GPX format}
\usage{
fm_export_gpx(input, name, outname)
}
\arguments{
\item{input}{a \code{sf} object to be saved as GPX.  _NOTE: currently only supports POINT objects}

\item{name}{Attribute to use to name the points.}

\item{outname}{Name of the output file}
}
\value{
a Boolean to confirm the successsful creation of the output file.
}
\description{
Takes an \code{sf} object and saves in to GPX format for use with GPS units.
}
