% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/forestr.R
\name{forestr}
\alias{forestr}
\alias{print.forestr}
\title{Classification and Regression with Random Forest and Choice of Splitting Function}
\usage{
forestr(formula, data, mvars, B = 500, min_size, method, ...)

\method{print}{forestr}(x, ...)
}
\arguments{
\item{formula,x}{formula specification of forest variables,
(for the \code{print} method, a \code{forestr} object)}

\item{data}{dataframe with y and x variables}

\item{mvars}{number of variables to sample in each node of
the tree during building the random forest}

\item{B}{number of bootstrap samples}

\item{min_size}{minimum size of a terminal node}

\item{method}{optional splitting method, currently \code{"extremes"} and \code{"purity"} implemented.}

\item{...}{extra parameters to pass to rpart}
}
\value{
An object of class \code{forestr} with components
  \item{call}{the original call to \code{forest}}
  \item{type}{one of \code{regression} or \code{classification}}
  \item{predicted}{the predicted values of the input data based on out of bag samples}
  \item{importance}{mean decrease in accurace over all classes for each variable in the model}
  \item{votes}{matrix giving the votes for each class on each observation (classification)}
  \item{oob}{out of bag error for the model including confusion matrix (classification)}
  \item{trees}{random forest tree objects}
}
\description{
Runs a random forest with user specified splitting criteria
using the \code{forestr} function.
}
\examples{
forestr(factor(cyl) ~ ., data = mtcars)
}

