% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transect_length.R
\name{get_transect_length}
\alias{get_transect_length}
\title{Get transect length of PCL transect (in meters)}
\usage{
get_transect_length(df, marker.spacing)
}
\arguments{
\item{df}{data frame of unprocessed PCL data}

\item{marker.spacing}{distance between transect markers, typically 10 m}
}
\value{
length of transect
}
\description{
\code{get_transect_length} imports and processes a single PCL transect.
}
\details{
This is a specific function that works using the input of a data directory
and a filename in that directory.Future versions of this will allow for direct
input of file into command so there will be no need to specify both directory and file.
}
\examples{

\dontrun{

}
}

