% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{higgs_10}
\alias{higgs_10}
\title{200 Simulated particles to emulate Higgs boson discovery}
\format{A data frame with 200 rows and 29 variables:
\describe{
   \item{class}{presence of Higgs boson particle, 1 for signal, 0 for background}
   \item{lepton_pT}{low- }
   \item{lepton_eta}{low-level feature }
   \item{lepton_phi}{low-level feature }
   \item{missing_energy_magnitude}{low-level feature }
   \item{missing_energy_phi}{low-level feature }
   \item{jet_1_pt}{low-level feature }
   \item{jet_1_eta}{low-level feature }
   \item{jet_1_phi}{low-level feature }
   \item{jet_1_b_tag}{low-level feature }
   \item{jet_2_pt}{low-level feature }
   \item{jet_2_eta}{low-level feature }
   \item{jet_2_phi}{low-level feature }
   \item{jet_2_b_tag}{low-level feature }
   \item{jet_3_pt}{low-level feature }
   \item{jet_3_eta}{low-level feature }
   \item{jet_3_phi}{low-level feature }
   \item{jet_3_b_tag}{low-level feature }
   \item{jet_4_pt}{low-level feature }
   \item{jet_4_eta}{low-level feature }
   \item{jet_4_phi}{low-level feature }
   \item{jet_4_b_tag}{low-level feature }
   \item{m_jj}{high-level feature }
   \item{m_jjj}{high-level feature }
   \item{m_lv}{high-level feature }
   \item{m_jlv}{high-level feature }
   \item{m_bb}{high-level feature }
   \item{m_wbb}{high-level feature }
   \item{m_wwbb}{high-level feature }
}}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/HIGGS}
}
\usage{
data(higgs_10)
}
\description{
A dataset containing simulated data produced using Monte Carlo
to resemble properties measured by particle detectors in an accelerator
to detect the Higgs Boson particle. This dataset created by sampling observations
to ensure the percentage of unbalancedness is 10%.
}
\keyword{datasets}

