% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idw-interval-value.R
\name{idw_interval_value}
\alias{idw_interval_value}
\title{IDW for especific window}
\usage{
idw_interval_value(ref_data, ref_key, ref_value, ref_lon, ref_lat, tgt_lon,
  tgt_lat, start, end, fun, max_dist = 200)
}
\description{
IDW for especific window
}
