% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{importance.forestr}
\alias{importance.forestr}
\title{Importance on a forestr object}
\usage{
\method{importance}{forestr}(object, ...)
}
\arguments{
\item{object}{an object of class forest as created by \code{forest}}

\item{...}{not currently used}
}
\value{
A data frame with the proximity values.
}
\description{
tidy on a NULL input returns an empty data frame, which means it can be
combined with other data frames (treated as "empty")
}

