% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{importance}
\alias{importance}
\title{Diagnostic methods for \code{forestr} to compute the importance for each variable}
\usage{
importance(object, ...)
}
\arguments{
\item{object}{an object of class forest as created by \code{forest}}

\item{...}{not currently used}
}
\value{
A data frame with the variable importance.
}
\description{
Diagnostic methods for \code{forestr} to compute the importance for each variable
}
\examples{
mtcars.forestr <- forestr(factor(cyl) ~ ., data = mtcars)
importance(mtcars.forestr)


}

