% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_matrix.R
\name{make_matrix}
\alias{make_matrix}
\title{Make PCL matrix for higher level complexity measures}
\usage{
make_matrix(df)
}
\arguments{
\item{df}{data frame of PCL data that has been processed with
\code{split_transect_from_pcl}}
}
\value{
sorted matrix of LiDAR returns for each x, z position
}
\description{
\code{make_matrix} produces a matrix of, x, z values in
coordinate space with the number and type of each LiDAR
return in each x, z bin combination
}
\details{
The \code{make_matrix} function munges data in to a data frame
of x, z bins with the number of canopy hits located in each bin.
}
\examples{
pcl_matrix <- make_matrix(pcl_split)

}
\keyword{matrix}
