% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_matrix.R
\name{make_matrix_part_one}
\alias{make_matrix_part_one}
\title{Make PCL matrix for higher level complexity measures}
\usage{
make_matrix_part_one(df)
}
\arguments{
\item{df}{data frame of unprocessed PCL data}
}
\value{
sorted matrix with z and x coordinates in meters
}
\description{
\code{make matrix} imports and processes a single PCL transect.
}
\examples{

\dontrun{
make_matrix(df)
}
}
\keyword{matrix}
