% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_summary_matrix.R
\name{make_summary_matrix}
\alias{make_summary_matrix}
\title{Creates summary matrix}
\usage{
make_summary_matrix(df, m)
}
\arguments{
\item{df}{sorted data frame of processed PCL data}

\item{m}{matrix of PCL hit density with x and z coordinates}
}
\value{
a matrix of summary stats by each x and z coordinate position
}
\description{
\code{make_summary_matrix} imports and processes a single PCL transect.
}
\details{
This makes a dataframe that is as long as a transect is. If the
transect is 40 m, this data frame has 40 rows.

#' This function allows you to express your love of cats.
}
\examples{

\dontrun{
#' make_summary_matrix()
}
}
\keyword{matrix}
\keyword{summary}
