% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_pcl.R
\name{normalize_pcl}
\alias{normalize_pcl}
\title{Normalize PCL data based on light saturation and attenuation}
\usage{
normalize_pcl(df)
}
\arguments{
\item{df}{data frame of pcl hit density}
}
\value{
a data frame of PCL hit density corrected for light saturation and attentuation based on Beer's Law
}
\description{
\code{normalize_pcl} imports and processes a single PCL transect.
}
\details{
This is a specific function that works using the input of a data directory
and a filename in that directory.Future versions of this will allow for direct
input of file into command so there will be no need to specify both directory and file.
}
\examples{
normalize_pcl()

\dontrun{

}
}
\keyword{light}

