% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_pcl_mh.R
\name{normalize_pcl_mh}
\alias{normalize_pcl_mh}
\title{Normalize PCL data based on light saturation and attenuation}
\usage{
normalize_pcl_mh(df, k)
}
\arguments{
\item{df}{data frame of pcl hit density processed from
\code{make_matrix}}

\item{k}{is a correction coefficent based on k = mean est. LAI/site LAI}
}
\value{
a data frame of PCL hit density corrected for
light saturation and attentuation based on Beer's Law
}
\description{
\code{normalize_pcl_mh} normalizes a PCL matrix for occlusion.
}
\details{
This function corrects saturated columns of LiDAR data
for occlusion based on assumptions from the Beer-Lambert Law.
}
\examples{
pcl_norm <- normalize_pcl_mh(pcl_matrix, k = 1)

}
\keyword{light}
