% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{osbs}
\alias{osbs}
\title{PCL transect from Ordway-Swisher Biological Station, Florida, US.}
\format{
A data frame with 10506 rows:
\describe{
  \item{index}{index of raw data--position along transect}
  \item{return_distance}{raw, uncorrected LiDAR return distances from laser}
  \item{intensity}{intensity values as recorded by LiDAR system}
}
}
\source{
\url{http://atkinsjeff.github.io}
}
\usage{
osbs
}
\description{
A dataset that consists of one 40 m transect taken in a longleaf pine-oak savanna
in North-central Florida. Data collected April, 2016 by J. Atkins and R. Fahey.
}
\keyword{datasets}
