% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcl_data}
\alias{pcl_data}
\title{PCL transect from the University of Virginia}
\format{An object of class \code{data.frame} with 14576 rows and 3 columns.}
\source{
\url{http://atkinsjeff.github.io}
}
\usage{
pcl_data
}
\description{
Derived from data collected at the University of Virginia
Data collected August, 2016 by J. Atkins. Dervied from the calc_vai function
}
\details{
#' @format A data frame with 14576rows:
\describe{
  \item{index}{index of raw data--position along transect}
  \item{return_distance}{raw, uncorrected LiDAR return distances from laser}
  \item{intensity}{intensity values as recorded by LiDAR system}
}
}
\keyword{datasets}
