% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcl_diagnostic_plot.R
\name{pcl_diagnostic_plot}
\alias{pcl_diagnostic_plot}
\title{PCL diagnostic plot}
\usage{
pcl_diagnostic_plot(df, filename)
}
\arguments{
\item{df}{data frame of unprocessed PCL data}

\item{filename}{name of file currently being processed}
}
\value{
a plot of PCL data showing marker spacing
}
\description{
\code{pcl_diagnostic_plot} imports and processes a single PCL transect.
}
\details{
This is a specific function that works using the input of a data directory
and a filename in that directory.Future versions of this will allow for direct
input of file into command so there will be no need to specify both directory and file.
}
\examples{

\dontrun{

}
}
