% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcl_matrix}
\alias{pcl_matrix}
\title{a LiDAR hit density by x, z position}
\format{A data frame with 1120 rows:
\describe{
  \item{xbin}{x-bin position}
  \item{zbin}{z-bin position}
  \item{bin.hits}{number of LiDAR returns at each x- and z- bin}
  \item{sky.hits}{total numer of sky hits per x column}
  \item{can.hits}{total numer of canopy hits per x column}
  \item{lidar.pulses}{no. of lidar pulses emitted per column}
  \item{Freq}{no idea}
  }
  @source \url{http://atkinsjeff.github.io}}
\usage{
pcl_matrix
}
\description{
Derived from data collected at the University of Virginia
Data collected August, 2016 by J. Atkins. Dervied from the calc_vai function
}
\keyword{datasets}
