% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcl_split}
\alias{pcl_split}
\title{a data frame LiDAR returns that have been split to x and z position}
\format{A data frame with 13982 rows:
\describe{
  \item{index}{index of raw data--position along transect}
  \item{return_distance}{raw, uncorrected LiDAR return distances from laser}
  \item{intensity}{intensity values as recorded by LiDAR system}
  \item{sky_hit}{lidar return that does not hit the canopy}
  \item{can_hit}{lidar return that hits the canopy}
  \item{marker}{negative value that indicates marker}
  \item{seg_num}{intermediate to get x position}
  \item{chunk_num}{intermediate to get x position}
  \item{xbin}{position along horizontal axis}
  \item{zbin}{position along vertical axis}
}}
\source{
\url{http://atkinsjeff.github.io}
}
\usage{
pcl_split
}
\description{
Derived from data collected at the University of Virginia
Data collected August, 2016 by J. Atkins. Dervied from the calc_vai function
}
\keyword{datasets}
