% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcl_summary}
\alias{pcl_summary}
\title{summary matrix}
\format{A data frame with 40 rows:
\describe{
  \item{xbin}{x-bin position}
  \item{mean.ht}{mean height}
  \item{sd.ht}{standard deviation of mean leaf height}
  \item{max.ht}{max measured height}
  \item{max.vai}{highest measured max VAI}
  \item{sum.vai}{total VAI for the column}
  \item{sd.vai}{standard deviation of VAI}
  \item{vai.z.sum}{density adjuste height}
  \item{max.vai.z}{height of peak VAI}
  \item{height.bin}{mean leaf height}
}}
\source{
\url{http://atkinsjeff.github.io}
}
\usage{
pcl_summary
}
\description{
Derived from data collected at the University of Virginia
Data collected August, 2016 by J. Atkins. Dervied from the calc_vai function
}
\keyword{datasets}
