% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hit_grid.R
\name{plot_hit_grid}
\alias{plot_hit_grid}
\title{Plots LiDAR hit grids of VAI}
\usage{
plot_hit_grid(m, filename, transect.length, max.ht, max.vai)
}
\arguments{
\item{m}{matrix of light adjusted vai values.}

\item{filename}{the name of the file currently being processed.}

\item{transect.length}{the length of the transect used to create the x-axis}

\item{max.ht}{the maximum measured height used to create the y-axis}

\item{max.vai}{the maximum density of VAI, defaul = 8}
}
\value{
a hit gride of VAI
}
\description{
\code{plot_hit_grid} produces a LiDAR hit grid plot
}
\examples{
# Calculates metrics of canopy structural complexity.
plot_hit_grid(pcl_vai, filename = "UVA LiDAR data", transect.length = 40,
 max.ht = 30, max.vai = 8)

}
\keyword{graphics}
