% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pavd.R
\name{plot_pavd}
\alias{plot_pavd}
\title{Graphs Plant Area Volume Density Profiles}
\usage{
plot_pavd(m, filename, plot.file.path.pavd, hist = FALSE)
}
\arguments{
\item{m}{matrix of light adjusted vai values.}

\item{filename}{the name of the file currently being processed.}

\item{plot.file.path.pavd}{path of plot file to be written, inherited from [process_pcl] or [process_multi_pcl]}

\item{hist}{logical input to include histogram of VAI, if TRUE it is included, if FALSE, it is not.}
}
\value{
plant area volume density plots
}
\description{
\code{plot_pavd} produces a PAVD plot from matrix data
}
\details{
This function is a nested function inside of [process_pcl]. It could be run independently using the summary_matrix.csv
output files created from running [procesS_pcl] as well.
}
\examples{
# Calculates metrics of canopy structural complexity.
\dontrun{plot_pavd(summary_matrix)
}
}
\keyword{area}
\keyword{density}
\keyword{pavd}
\keyword{plant}
\keyword{profile,}
\keyword{volume}
