% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\alias{predict.forestr}
\title{Predict}
\usage{
\method{predict}{forestr}(object, newdata, ...)
}
\arguments{
\item{object}{an object of class forest as created by \code{forest}}

\item{newdata}{dataframe containing new data}

\item{...}{not currently used}
}
\value{
currently nothing
}
\description{
Predict methods to predict from forestr tree and random forest
within forestr using the \code{predict} function
}
\examples{
mtcars.forest <- forest(factor(cyl) ~ ., data = mtcars[1:25, ])
predict(mtcars.forest, mtcars[26:32, ])
}

