% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pcl_transect.R
\name{process_pcl_transect}
\alias{process_pcl_transect}
\title{Process single PCL transects.}
\usage{
process_pcl_transect(data_dir, filename)
}
\description{
\code{process_pcl_transect} imports and processes a single PCL transect.
}
\details{
This is a specific function that works using the input of a data directory
and a filename in that directory.Future versions of this will allow for direct
input of file into command so there will be no need to specify both directory and file.
}
\examples{
data_directory <- "./data/PCL_transects/"  #data directory containing PCL transects
filename <- "oldgrowth_one.csv"  #name of PCL transect to be processed
process_pcl_transect(data_dir, filename)

process_pcl_transect("./data/PCL_transects/", "oldgrowth_one.csv" )

sum(.Machine$integer.max, 1L)
sum(.Machine$integer.max, 1)

\dontrun{

}
}

