% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_tls.R
\name{process_tls}
\alias{process_tls}
\title{Process single PCL transects.}
\usage{
process_tls(f, slice, pavd = FALSE, hist = FALSE)
}
\arguments{
\item{f}{the name of the filename to input <character> or a data frame <data frame>.}

\item{slice}{the number of the transect to use from xyz tls data}

\item{pavd}{logical input to include Plant Area Volume Density Plot from [plot_pavd], if TRUE it is included, if FALSE, it is not.}

\item{hist}{logical input to include histogram of VAI with PAVD plot, if TRUE it is included, if FALSE, it is not.}
}
\value{
writes the hit matrix, summary matrix, and output variables to csv in an output folder, along with hit grid plot
}
\description{
\code{process_tls} imports and processes a slice from a voxelated TLS scan.
}
\details{
This function takes as input a four column .CSV file or data frame of x, y, z, and VAI
(Vegetation Area Index) derived from 3-D LiDAR data. Currently, this function only analyzes
a single slice from the inputed TLS data set. VAI is calculated externally by the user using
user-determined methodologyy.
}
\examples{

# with designated file
\dontrun{process_tls("tls.csv", slice = 5, pavd = FALSE, hist = FALSE)
}

}
\keyword{processing}
\keyword{tls}
