% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{proximity.forestr}
\alias{proximity.forestr}
\title{Proximity on a forestr object}
\usage{
\method{proximity}{forestr}(object, ...)
}
\arguments{
\item{object}{an object of class forest as created by \code{forest}}

\item{...}{not currently used}
}
\value{
A data frame with the proximity values.
}
\description{
tidy on a NULL input returns an empty data frame, which means it can be
combined with other data frames (treated as "empty")
}

