% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{proximity}
\alias{proximity}
\title{Diagnostic methods for \code{forestr} to compute the proximity matrix}
\usage{
proximity(object, ...)
}
\arguments{
\item{object}{an object of class forest as created by \code{forest}}

\item{...}{not currently used}
}
\value{
A data frame with the proximity values.
}
\description{
Diagnostic methods for \code{forestr} to compute the proximity matrix
}
\examples{
mtcars.forestr <- forestr(factor(cyl) ~ ., data = mtcars)
proximity(mtcars.forestr)
}

