% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{red_pine}
\alias{red_pine}
\title{PCL transect from a red pine plantation in Northern Michigan, US.}
\format{A data frame with 17559 rows:
\describe{
  \item{index}{index of raw data--position along transect}
  \item{return_distance}{raw, uncorrected LiDAR return distances from laser}
  \item{intensity}{intensity values as recorded by LiDAR system}
}}
\source{
\url{http://atkinsjeff.github.io}
}
\usage{
red_pine
}
\description{
A dataset that consists of one 40 m transect taken in a red pine plantations
in Northern Michigan. Data collected July, 2017 by J. Atkins.
}
\keyword{datasets}
