% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ-var.R
\name{summ_var_part}
\alias{summ_var_part}
\title{Summarise value by groups and minimum observation}
\usage{
summ_var_part(key, vec, fun = mean, groups = 3, n_min = 6, ...)
}
\description{
Summarise value by groups and minimum observation
}
