% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_hit_matrix_to_csv.R
\name{write_hit_matrix_to_csv}
\alias{write_hit_matrix_to_csv}
\title{Writes hit matrix to csv for further analysis}
\usage{
write_hit_matrix_to_csv(m, filename, output_directory)
}
\arguments{
\item{m}{matrix of VAI with z and x coordinates}

\item{filename}{name of file currently being processed}
}
\description{
\code{write_hit_matrix_to_csv} writes hit matrix to .csv for further analysis
}
\details{
This is a specific function that writes the output variables to disk in .csv format
}
\examples{
write_hit_matrix_to_csv)

}
\keyword{hit}
\keyword{matrix}

