% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_summary_matrix_to_csv.R
\name{write_summary_matrix_to_csv}
\alias{write_summary_matrix_to_csv}
\title{Writes csc metrics and output variables to .csv}
\usage{
write_summary_matrix_to_csv(m, filename)
}
\arguments{
\item{m}{summary matrix}

\item{filename}{name of file currently being processed}
}
\description{
\code{write_summary_matrix_to_csv} writes summary matrix to .csv format
}
\details{
This is a specific function that writes the summary matrix to disk in .csv format
}
\examples{
write_summary_matrix_to_csv()

}
\keyword{matrix}
\keyword{summary}

