% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baselayers}
\alias{baselayers}
\title{Baselayers}
\format{
An object of class \code{data.frame} with 10 rows and 4 columns.
}
\source{
Data layers from the BC Data Catalog.  See the package "bcdata"
}
\usage{
baselayers
}
\description{
A dataframe BC Data Catalog layers to queried https://catalogue.data.gov.bc.ca/.  Includes a human readable name and the BC Data ID.
}
\keyword{datasets}
