% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_enl.R
\name{calc_enl}
\alias{calc_enl}
\title{Calculate rugosity and other higher level complexity metrics}
\usage{
calc_enl(m)
}
\arguments{
\item{m}{the summary matrix}
}
\value{
a slew of metrics...need to fully outline later
}
\description{
\code{calc_enl} imports and processes a single PCL transect.
}
\details{
This is a specific function calculates canopy rugosity and other metrics
}
\examples{
# Calculates the effective number of layers
\dontrun{calc_enl()
}
}
\keyword{rugosity}
