% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rumple.R
\name{calc_rumple}
\alias{calc_rumple}
\title{Calculates rumple}
\usage{
calc_rumple(df)
}
\arguments{
\item{df}{LiDAR summary matrix data frame}
}
\value{
rumple for the canopy based on 2-D transect
}
\description{
\code{calc_rumple} calculates canopy rumple.
}
\details{
This function uses the summary matrix created by
the function \code{make_summary_matrix} to calculate
canopy rumple, the relationship between outer canopy surface
and the ground area.
}
\examples{

calc_rumple(pcl_summary)

}
\keyword{rumple}
