% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tls_csc.R
\name{calc_tls_csc}
\alias{calc_tls_csc}
\title{Calculates rumple}
\usage{
calc_tls_csc(m, filename)
}
\arguments{
\item{m}{matrix of vai data with mean leaf height column}

\item{filename}{the name of the file being process0}
}
\value{
csc metrics
}
\description{
\code{calc_tls_csc} calculates canopy structural complexity metrics from the tls vai matrix
}
\details{
This is a specific function to calculate canopy structural complexity or CSC metrics from the VAI matrix imported in.
}
\examples{
\dontrun{
calc_tls_csc(m)
}
}
\keyword{csc}
\keyword{rugosity}
\keyword{tls}
