% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_hits.R
\name{code_hits}
\alias{code_hits}
\title{Code hits}
\usage{
code_hits(df)
}
\arguments{
\item{df}{a raw set of pcl data}
}
\description{
\code{code_hits} classifies data values as canopy returns, sky returns, or
data markers.
}
\details{
The function \code{code_hits} accounts for the NAs that are in
 the return distance column which are actually
  the sky hits (i.e. when the lidar does not record a canopy hit).
}
\examples{
# classify data values that have been imported using read_pcl
pcl_coded <- code_hits(pcl_data)

}
