% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_variables.R
\name{combine_variables}
\alias{combine_variables}
\title{Calculate rugosity and other higher level complexity metrics}
\usage{
combine_variables(
  variable.list,
  csc.metrics,
  rumple,
  clumping.index,
  enl,
  intensity_stats
)
}
\arguments{
\item{variable.list}{a list of variables from calc_rugosity}

\item{csc.metrics}{a list of metrics from the csc_metrics function}

\item{rumple}{rumple value from calc_rumple}

\item{clumping.index}{clumping index value}

\item{enl}{effective number of layers}

\item{intensity_stats}{statistics on intensity values}
}
\value{
a concatenated list of csc variables
}
\description{
\code{combine_variables} combines varibles from internal functions to export
}
\details{
This is a specific function combines variables together to a coherent list.
}
\examples{
\dontrun{
#This function runs interanlly and collates the slew of
CSC metrics and formats them.
#' combine_Variables(variable.list, csc.metrics, rumple, clumping.index)
}
}
\keyword{csc}
