% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csc_metrics.R
\name{csc_metrics}
\alias{csc_metrics}
\title{Cover and sky fraction estimates}
\usage{
csc_metrics(df, filename)
}
\arguments{
\item{df}{data frame of uncorrected PCL data}

\item{filename}{name of file currently being processed}
}
\value{
slew of cover and sky fraction metrics
}
\description{
\code{csc_metrics} imports and processes a single PCL transect.
}
\details{
This is a specific function that works using the input of a data directory
and a filename or a single data frame in that directory.
}
\examples{

\dontrun{

}
}
\keyword{canopy}
\keyword{csc,}
\keyword{fraction}
\keyword{fraction,}
\keyword{rugosity,}
\keyword{sky}
\keyword{structure,}
