% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_aoi.R
\name{fm_aoi}
\alias{fm_aoi}
\title{Create an area of interest}
\usage{
fm_aoi(extent = bc, basemap = "Esri.WorldImagery", epsg = 3005)
}
\arguments{
\item{extent}{An sf object to define the extent}

\item{basemap}{a Leaflet basemap. Default is . Other providers include: \code{CartoDB.Positron}, \code{Esri.WorldImagery}, \code{Esri.DeLorme}.  See also: https://leaflet-extras.github.io/leaflet-providers/preview/}

\item{epsg}{EPSG code for returned projection information.  Defaults to BC Albers (epsg 3005)}
}
\value{
Returns an sf object.
}
\description{
Creates an area of interest interactively.  Simply call the function
}
\examples{
aoi <- fm_aoi(extent = bc, epsg = 3005)


}
