% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/forest.R
\name{forest}
\alias{forest}
\title{Forest}
\usage{
forest(formula, data, mvars, B = 1000)
}
\arguments{
\item{formula}{formula specification of forest variables}

\item{data}{dataframe with y and x variables}

\item{B}{number of bootstrap samples}
}
\value{
A data.frame with columns
  \item{b}{1:B}
  \item{sample}{Bootstrap sampled datasets.}
  \item{rf}{Random forest object}
}
\description{
Runs a random forest with user specified splitting criteria
using the \code{forest} function.
}
\examples{
forest(factor(cyl) ~ ., data = mtcars)
}

