% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_matrix.R
\name{make_matrix_part_one}
\alias{make_matrix_part_one}
\title{Make PCL matrix for higher level complexity measures}
\usage{
make_matrix_part_one(df)
}
\arguments{
\item{df}{data frame of unprocessed PCL data}
}
\value{
sorted matrix with z and x coordinates in meters
}
\description{
\code{make matrix} imports and processes a single PCL transect.
}
\examples{
data_directory <- "./data/PCL_transects/"  #data directory containing PCL transects
filename <- "oldgrowth_one.csv"  #name of PCL transect to be processed
process_pcl(data_directory, filename)

process_pcl("./data/PCL_transects/", "oldgrowth_one.csv" )

\dontrun{

}
}
\keyword{matrix}

