% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_matrix_part_one.R
\name{make_matrix_part_one}
\alias{make_matrix_part_one}
\title{Make PCL matrix part one}
\usage{
make_matrix_part_one(df)
}
\arguments{
\item{df}{data frame of PCL data that has been processed with}
}
\value{
sorted matrix of LiDAR returns for each x, z position
}
\description{
\code{make_matrix_part_one} produces a matrix of, x, z values in
coordinate space with the number and type of each LiDAR
return in each x, z bin combination
}
\keyword{matrix}
