% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_summary_matrix.R
\name{make_summary_matrix}
\alias{make_summary_matrix}
\title{Creates summary matrix}
\usage{
make_summary_matrix(df, m)
}
\arguments{
\item{df}{sorted data frame of processed PCL data}

\item{m}{matrix of PCL hit density with x and z coordinates}
}
\value{
a matrix of summary stats by each x and z coordinate position
}
\description{
\code{make_summary_matrix} creates a summary matrix of data through data wrangling
the VAI data frame.
}
\details{
This makes a dataframe that is as long as a transect is. If the
transect is 40 m, this data frame has 40 rows. As input,
\code{make_summary_matrix} requires a data frame of values
from \code{split_transects_from_pcl} first, and second,
the data frame of VAI from the function \code{calc_vai}.

#' This function allows you to express your love of cats.
}
\examples{

pcl_summary <- make_summary_matrix(pcl_split, pcl_vai)

}
\keyword{matrix}
\keyword{summary}
