% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_pcl.R
\name{normalize_pcl}
\alias{normalize_pcl}
\title{Normalize PCL data based on light saturation and attenuation}
\usage{
normalize_pcl(df)
}
\arguments{
\item{df}{data frame of pcl hit density processed from
\code{make_matrix}}
}
\value{
a data frame of PCL hit density corrected for
light saturation and attentuation based on Beer's Law
}
\description{
\code{normalize_pcl} normalizes a PCL matrix for occlusion.
}
\details{
This function corrects saturated columns of LiDAR data
for occlusion based on assumptions from the Beer-Lambert Law.
}
\examples{
pcl_norm <- normalize_pcl(pcl_matrix)

}
\keyword{light}
