% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcl_diagnostic_plot.R
\name{pcl_diagnostic_plot}
\alias{pcl_diagnostic_plot}
\title{PCL diagnostic plot}
\usage{
pcl_diagnostic_plot(df, filename)
}
\arguments{
\item{df}{data frame of unprocessed PCL data}

\item{filename}{name of file currently being processed}
}
\value{
a plot of PCL data showing marker spacing
}
\description{
\code{pcl_diagnostic_plot} this function provides a diagnostic view of raw PCL data
}
\details{
This function provides a graphic view of raw PCL data to check for equal data
spacing and marker spacing
}
\examples{
# using the Ordway-Swisher Data set
pcl_diagnostic_plot(osbs)

}
