% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pavd.R
\name{plot_pavd}
\alias{plot_pavd}
\title{Graphs Plant Area Volume Density Profiles}
\usage{
plot_pavd(m, filename, plot.file.path.pavd, hist = FALSE)
}
\arguments{
\item{m}{matrix of light adjusted vai values.}

\item{filename}{the name of the file currently being processed.}

\item{plot.file.path}{path of plot file to be written, inherited from [process_pcl] or [process_multi_pcl]}
}
\value{
plant area volume density plots
}
\description{
\code{plot_pavd} produces a PAVD plot from matrix data
}
\details{
This is a specific function calculates canopy rugosity and other metrics
}
\examples{
# Calculates metrics of canopy structural complexity.
\dontrun{plot_pavd()
}
}
\keyword{area}
\keyword{density}
\keyword{pavd}
\keyword{plant}
\keyword{profile,}
\keyword{volume}
