% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_multi_pcl.R
\name{process_multi_pcl}
\alias{process_multi_pcl}
\title{Process multiplie PCL transects.}
\usage{
process_multi_pcl(data_dir, user_height, marker.spacing, max.vai)
}
\arguments{
\item{data_dir}{directory where PCL .csv files are stored}

\item{user_height}{height of laser from ground based on user in meters}

\item{marker.spacing}{space between markers in the PCL data, in meters}

\item{max.vai}{the maximum value of column VAI. The default is 8. Should be a max value, not a mean.}
}
\description{
\code{process_multi_pcl} imports and processes mutiple PCL transect.
}
\details{
This is a specific function that works using the input of a data directory of .csv
files where the function cycles through the files there.
}
\examples{
\dontrun{
# This function works on a directory of raw PCL data
data_directory <- "./data/PCL_transects/"  #data directory containing PCL transects
process_multi_pcl(data_directory)

process_multi_pcl("./data/PCL_transects/"
}

}
