% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pcl.R
\name{process_pcl}
\alias{process_pcl}
\title{Process single PCL transects.}
\usage{
process_pcl(f, user_height, marker.spacing)
}
\arguments{
\item{f}{the name of the filename to input}

\item{user_height}{the height of the laser off the ground as mounted on the user in meters}

\item{marker.spacing}{distance between markers, typically 10 m}
}
\value{
writes the hit matrix, summary matrix, and output variables to csv in an output folder, along with hit grid plot
}
\description{
\code{process_pcl} imports and processes a single PCL transect.
}
\details{
This is a specific function that works using the input of a data directory
and a filename in that directory.Future versions of this will allow for direct
input of file into command so there will be no need to specify both directory and file.
}
\examples{
data_directory <- "./data/PCL_transects/"  #data directory containing PCL transects
filename <- "oldgrowth_one.csv"  #name of PCL transect to be processed
process_pcl(data_directory, filename)

process_pcl("./data/PCL_transects/", "oldgrowth_one.csv" )


\dontrun{

}
}
\keyword{pcl}
\keyword{processing}

