% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_pcl.R
\name{process_pcl}
\alias{process_pcl}
\title{Process single PCL transects.}
\usage{
process_pcl(f, user_height, marker.spacing, max.vai)
}
\arguments{
\item{f}{the name of the filename to input <character> or a data frame <data frame>.}

\item{user_height}{the height of the laser off the ground as mounted on the user in meters. default is 1 m}

\item{marker.spacing}{distance between markers, defaults is 10 m}

\item{max.vai}{the maximum value of column VAI. The default is 8. Should be a max value, not a mean.}
}
\value{
writes the hit matrix, summary matrix, and output variables to csv in an output folder, along with hit grid plot
}
\description{
\code{process_pcl} imports and processes a single PCL transect.
}
\details{
This is function works on either files or existing data frames in the environment. It processes raw PCL data
through a workflow that cuts the data into 1 meter segments with z and x positions and vertically normalizes data based on
light extinction assumptions from the Beer-Lambert Law to account for light saturation. Data are then
summarized, metrics of canopy structure complexity are calculated, and then output data saved to an output directory.
A hit grid plot is also saved to this same directory.
}
\examples{

# with designated file
\dontrun{process_pcl("pcl_data.csv", marker.spacing = 10, user_height = 1.05, max.vai = 8)


# with data frame
process_pcl(osbs, marker.spacing = 10, user_height = 1.05, max.vai = 8)
}
}
\keyword{pcl}
\keyword{processing}
