% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_transects_from_pcl.R
\name{split_transects_from_pcl}
\alias{split_transects_from_pcl}
\title{Split transects from PCL}
\usage{
split_transects_from_pcl(pcl_data, transect.length, marker.spacing,
  DEBUG = FALSE, write_out = FALSE, data_dir, output_file_name)
}
\arguments{
\item{pcl_data}{data frame of unprocessed PCL data}

\item{transect.length}{total transect length}

\item{marker.spacing}{distance between markers in meters within the PCL data}

\item{data_dir}{directory where PCL data .csv are stored}

\item{output_file_name}{old code relic}
}
\description{
\code{split_transects_from_pcl} imports and processes a single PCL transect.
}
\details{
Function to add two additional columns
to the pcl dataset, one for the segment
 (which should only be from 1-4) and is
designated by a -99999999 value in the
return_distance column
The only required parameter is the data
frame of pcl data, but this can
optionally also write out the results
to csv if a path and name are given
}
\examples{

\dontrun{

}
}
