% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_hit_matrix_to_csv.R
\name{write_hit_matrix_to_csv}
\alias{write_hit_matrix_to_csv}
\title{Writes hit matrix to csv for further analysis}
\usage{
write_hit_matrix_to_csv(m, outputname, output_directory)
}
\arguments{
\item{m}{matrix of VAI with z and x coordinates}

\item{outputname}{name of file currently being processed}

\item{output_directory}{directory where output goes}
}
\description{
\code{write_hit_matrix_to_csv} writes hit matrix to .csv for further analysis
}
\details{
This is a specific sub-function that writes the output variables to disk in .csv format
and runs within the functions \code{process_pcl}, \code{process_multi_pcl}, and
\code{proces_tls}.
}
\examples{
\dontrun{
# This function runs internally.
write_hit_matrix_to_csv(m, outputname, output_directory)
}
}
\seealso{
\code{\link{process_pcl}}
\code{\link{write_pcl_to_csv}}
\code{\link{write_summary_matrix_to_csv}}
}
\keyword{hit}
\keyword{matrix}
