% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pcl_to_csv.R
\name{write_pcl_to_csv}
\alias{write_pcl_to_csv}
\title{Writes csc metrics and output variables to .csv}
\usage{
write_pcl_to_csv(output.variables, outputname, output_directory)
}
\arguments{
\item{output.variables}{list of concatenated output variables}

\item{outputname}{name of file currently being processed}

\item{output_directory}{directory where output goes}
}
\description{
\code{write_pcl_to_csv} writes csc metrics and varialbes to .csv format
}
\details{
This is a specific function that writes the output variables to disk in .csv format
and runs within the functions \code{process_pcl}, \code{process_multi_pcl}, and
\code{proces_tls}.
}
\examples{
\dontrun{
write_pcl_to_csv(output_variables, outputname, output_directory)
}

}
\seealso{
\code{\link{process_pcl}}
\code{\link{write_summary_matrix_to_csv}}
\code{\link{write_hit_matrix_to_csv}}
}
\keyword{output}
\keyword{variables}
