% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_summary_matrix_to_csv.R
\name{write_summary_matrix_to_csv}
\alias{write_summary_matrix_to_csv}
\title{Writes csc metrics and output variables to .csv}
\usage{
write_summary_matrix_to_csv(m, outputname, output_directory)
}
\arguments{
\item{m}{summary matrix}

\item{outputname}{name of file currently being processed}

\item{output_directory}{directory where output goes}
}
\description{
\code{write_summary_matrix_to_csv} writes summary matrix to .csv format
}
\details{
This is a specific subfunction that writes the summary matrix to disk in .csv format
and runs within the functions \code{process_pcl}, \code{process_multi_pcl}, and
\code{proces_tls}.
}
\examples{
\dontrun{
write_summary_matrix_to_csv()
}
}
\seealso{
\code{\link{write_pcl_to_csv}}
\code{\link{write_hit_matrix_to_csv}}
}
\keyword{matrix}
\keyword{summary}
